requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-multipart
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]